+++ portal-ui/src/app/components/OrgActions/Del/OrgActions-Del.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { action, makeObservable, observable } from 'mobx';
 import { observer } from 'mobx-react';
-import { Dialog, DialogActions, DialogContent, DialogContentText, IconButton, Tooltip } from '@mui/material';
+import { IconButton, Tooltip } from '@mui/material';
 import { Delete, DeleteOutline } from '@mui/icons-material';
 import { cn } from '@bem-react/classname';
 import { boundMethod } from 'autobind-decorator';
@@ -10,7 +10,7 @@ import { CrgGroup } from '../../../services/auth/groups/groups.models';
 import { groupsService } from '../../../services/auth/groups/groups.service';
 import { CrgUser } from '../../../services/auth/users/users.models';
 import { usersService } from '../../../services/auth/users/users.service';
-import { Button } from '../../Button/Button';
+import { konfirmieren } from '../../../services/utility-dialogs.service';
 import { Loading } from '../../Loading/Loading';
 
 const cnOrgActionsDel = cn('OrgActions', 'Del');
@@ -34,55 +34,57 @@ export class OrgActionsDel extends Component<OrgActionsDelProps> {
     return (
       <>
         <Tooltip title='Удалить'>
-          <IconButton className={cnOrgActionsDel()} onClick={this.openDialog}>
+          <IconButton className={cnOrgActionsDel()} onClick={this.handleDelete}>
             {this.dialogOpen ? <Delete /> : <DeleteOutline />}
           </IconButton>
         </Tooltip>
-        <Dialog open={this.dialogOpen} onClose={this.closeDialog}>
-          <DialogContent>
-            <DialogContentText>Удалить?</DialogContentText>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={this.delete} color='primary'>
-              Удалить
-            </Button>
-            <Button onClick={this.closeDialog}>Отмена</Button>
-          </DialogActions>
-        </Dialog>
         <Loading visible={this.loading} global />
       </>
     );
   }
 
-  @action.bound
-  private openDialog() {
-    this.dialogOpen = true;
-  }
-
-  @action.bound
-  private closeDialog() {
-    this.dialogOpen = false;
-  }
-
   @boundMethod
-  private async delete() {
+  private async handleDelete() {
+    this.setDialogOpen(true);
+    const shouldDelete = await konfirmieren({
+      title: 'Подтверждение удаления',
+      message: this.props.user
+        ? 'Вы уверены, что хотите удалить этого пользователя?'
+        : 'Вы уверены, что хотите удалить эту группу?',
+      okText: 'Удалить',
+      cancelText: 'Отмена'
+    });
+
+    if (!shouldDelete) {
+      this.setDialogOpen(false);
+
+      return;
+    }
+
     const { user, group } = this.props;
     this.setLoading(true);
-    this.closeDialog();
 
-    if (user) {
-      await usersService.delete(user);
+    try {
+      if (user) {
+        await usersService.delete(user);
+      }
+
+      if (group) {
+        await groupsService.delete(group);
+      }
+    } finally {
+      this.setLoading(false);
+      this.setDialogOpen(false);
     }
-
-    if (group) {
-      await groupsService.delete(group);
-    }
-
-    this.setLoading(false);
   }
 
   @action
   private setLoading(loading: boolean) {
     this.loading = loading;
+  }
+
+  @action
+  private setDialogOpen(open: boolean) {
+    this.dialogOpen = open;
   }
 }
